Evil Monkey Productions
A Narrative Fiction company with pending release of the unofficial and final installment of Legend Entertainment’s Spellcasting series. We are targeting a release date on schedule with the IF Comp, 2024, 30th Anniversary. Donations will be accepted once the climaxing story of Ernie and Lola is published. As always, thank you for your support.

evilmonkeyprod@gmail.com

Interactive Fiction Forum